new AtomicOperation module to portably capture InterlockedIncrement() functionality. Vector all existing calls to interlocked increment/decremnt/add/subtract through the new wrappers.

+++ Library/Sources/Stroika/Foundation/Memory/MemoryAllocator.h
@@ -123,8 +123,8 @@ namespace	Stroika {
 
 				private:
 					AbstractGeneralPurposeAllocator&	fBaseAllocator;
-					LONG								fNetAllocationCount;
-					LONG								fNetAllocatedByteCount;
+					uint32_t							fNetAllocationCount;
+					size_t								fNetAllocatedByteCount;
 			};
